Выносит временные данные в файл data.js
